+++ src/lib/lforms/lhc-form-data.ts
@@ -124,6 +124,8 @@ export default class LhcFormData {
   _fhirVariables;
   _hasResponsiveExpr;
   copyrightNotice;
+  url;
+  version;
   type;
   _showErrors;
   _showWarnings;
@@ -1431,6 +1433,8 @@ export default class LhcFormData {
       type: this.type,
       template: this.template,
       copyrightNotice: this.copyrightNotice,
+      url: this.url,
+      version: this.version,
       items: formData.itemsData,
       templateOptions: CommonUtils.deepCopy(this.templateOptions)
     };